init_createを再度作成
