+++ alembic/versions/6f2a56d43b94_init_create.py
@@ -1,0 +1,28 @@
+"""init_create
+
+Revision ID: 6f2a56d43b94
+Revises: 
+Create Date: 2023-08-07 12:31:22.826184
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '6f2a56d43b94'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
