--- alembic/versions/c0263f9526f2_init_create.py
@@ -1,77 +1,0 @@
-"""init create
-
-Revision ID: c0263f9526f2
-Revises: 
-Create Date: 2023-07-02 14:53:42.334165
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'c0263f9526f2'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        'users',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('name', sa.String(), nullable=True),
-        sa.Column('email', sa.String(), nullable=True),
-        sa.Column('password', sa.String(), nullable=True),
-        sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_index(op.f('ix_users_name'), 'users', ['name'], unique=False)
-    op.create_index(op.f('ix_users_password'), 'users',
-                    ['password'], unique=False)
-    op.create_table('tasks',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('user_id', sa.Integer(), nullable=True),
-                    sa.Column('name', sa.String(), nullable=True),
-                    sa.Column('order', sa.Integer(), nullable=True),
-                    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    op.create_index(op.f('ix_tasks_id'), 'tasks', ['id'], unique=False)
-    op.create_index(op.f('ix_tasks_name'), 'tasks', ['name'], unique=False)
-    op.create_index(op.f('ix_tasks_order'), 'tasks', ['order'], unique=True)
-    op.create_index(op.f('ix_tasks_user_id'), 'tasks',
-                    ['user_id'], unique=False)
-    op.create_table('rundates',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('task_id', sa.Integer(), nullable=True),
-                    sa.Column('date', sa.DateTime(), nullable=True),
-                    sa.ForeignKeyConstraint(['task_id'], ['tasks.id'], ),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    op.create_index(op.f('ix_rundates_date'),
-                    'rundates', ['date'], unique=False)
-    op.create_index(op.f('ix_rundates_id'), 'rundates', ['id'], unique=False)
-    op.create_index(op.f('ix_rundates_task_id'),
-                    'rundates', ['task_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_rundates_task_id'), table_name='rundates')
-    op.drop_index(op.f('ix_rundates_id'), table_name='rundates')
-    op.drop_index(op.f('ix_rundates_date'), table_name='rundates')
-    op.drop_table('rundates')
-    op.drop_index(op.f('ix_tasks_user_id'), table_name='tasks')
-    op.drop_index(op.f('ix_tasks_order'), table_name='tasks')
-    op.drop_index(op.f('ix_tasks_name'), table_name='tasks')
-    op.drop_index(op.f('ix_tasks_id'), table_name='tasks')
-    op.drop_table('tasks')
-    op.drop_index(op.f('ix_users_password'), table_name='users')
-    op.drop_index(op.f('ix_users_name'), table_name='users')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_table('users')
-    # ### end Alembic commands ###
